requirements.txt
redis==4.4.2
elasticsearch[async]==7.9.1
fastapi==0.99.1
orjson==3.9.2
python-dotenv==1.0.0
pydantic==1.9.0
uvicorn==0.12.2
gunicorn==20.1.0
httptools==0.6.0
python-jose
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-jaeger
opentelemetry-exporter-otlp
slowapi==0.1.8
uvloop==0.17.0 ; sys_platform != "win32" and implementation_name == "cpython"